requirements.txt
torch
torchvision
torchaudio
ultralytics
gradio
opencv-python
pillow
numpy
PySide6
pyyaml
PyQt5
